backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.135.3",
    "motor>=3.7.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "uvicorn[standard]>=0.43.0",
    "yfinance>=1.2.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
